### 2025-07-23 #298 @editorialbot
[PRE REVIEW]: BayesCog: A freely available course in Bayesian statistics and hierarchical Bayesian modeling for psychological science
**Submitting author:** @sohaamir (<a href="http://orcid.org/0009-0000-6584-4579">Aamir Sohail</a>)
**Repository:** https://github.com/Alpn-Lab/BayesCog
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
