### 2025-09-27 #298 @sohaamir
Hi @labarba, is there a way for me to assign specific reviewers and editor as I have observed in other submissions? I would have also done so originally when submitting but did not see the option.

i.e., instead of:

Editor: Pending
Reviewers: Pending

We would have (for example):

Editor: [at]JasonJWilliamsNY
Reviewers: [at]TomDonoghue, [at]arokem
