+++ paper_psyarxiv.bib
@@ -1,0 +1,451 @@
+@article{Ahn2017,
+  title={Revealing {Neurocomputational} {Mechanisms} of {Reinforcement} {Learning} and {Decision}-{Making} {With} the {hBayesDM} {Package}},
+  author={Ahn, Woo-Young and Haines, Nathaniel and Zhang, Lei},
+  journal={Computational Psychiatry (Cambridge, Mass.)},
+  volume={1},
+  pages={24--57},
+  year={2017}
+}
+
+@article{Annis2018,
+  title={Bayesian statistical approaches to evaluating cognitive models},
+  author={Annis, Jeffrey and Palmeri, Thomas J.},
+  journal={Wiley Interdisciplinary Reviews: Cognitive Science},
+  volume={9},
+  number={2},
+  pages={e1458},
+  year={2018},
+  publisher={Wiley Online Library}
+}
+
+@article{Baribault2023,
+  title={Troubleshooting {Bayesian} cognitive models},
+  author={Baribault, Beth and Collins, Anne G. E.},
+  journal={Psychological Methods},
+  year={2023}
+}
+
+@article{Carpenter2017,
+  title={Stan: {A} {Probabilistic} {Programming} {Language}},
+  author={Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D. and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus A. and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+  journal={Journal of Statistical Software},
+  volume={76},
+  pages={1},
+  year={2017}
+}
+
+@misc{Chang2020,
+  title={DartBrains: An online open access resource for learning functional neuroimaging analysis methods in Python},
+  author={Chang, Luke J. and Huckins, Jeremy and Cheong, Jin Hyun and Brietzke, Samuele and Lindquist, Martin A. and Wager, Tor D.},
+  year={2020},
+  publisher={Zenodo},
+  note={https://doi.org/10.5281/zenodo}
+}
+
+@article{Crawley2020,
+  title={Modeling flexible behavior in childhood to adulthood shows age-dependent learning mechanisms and less optimal learning in autism in each age group},
+  author={Crawley, Daisy and Zhang, Lei and Jones, Emily J. H. and Ahmad, Jumana and Oakley, Bethany and San Jos{\'e} C{\'a}ceres, Antonia and Charman, Tony and Buitelaar, Jan K. and Murphy, Declan G. M. and Chatham, Christopher and others},
+  journal={PLoS Biology},
+  volume={18},
+  number={10},
+  pages={e3000908},
+  year={2020},
+  publisher={Public Library of Science}
+}
+
+@article{Daw2011,
+  title={Model-based influences on humans' choices and striatal prediction errors},
+  author={Daw, Nathaniel D. and Gershman, Samuel J. and Seymour, Ben and Dayan, Peter and Dolan, Raymond J.},
+  journal={Neuron},
+  volume={69},
+  number={6},
+  pages={1204--1215},
+  year={2011},
+  publisher={Elsevier}
+}
+
+@article{Dayan2008,
+  title={Reinforcement learning: the good, the bad and the ugly},
+  author={Dayan, Peter and Niv, Yael},
+  journal={Current Opinion in Neurobiology},
+  volume={18},
+  number={2},
+  pages={185--196},
+  year={2008},
+  publisher={Elsevier}
+}
+
+@article{deHollander2016,
+  title={Different {Ways} of {Linking} {Behavioral} and {Neural} {Data} via {Computational} {Cognitive} {Models}},
+  author={de Hollander, Gilles and Forstmann, Birte U. and Brown, Scott D.},
[…]
