+++ paper_psyarxiv.md
@@ -1,0 +1,123 @@
+---
+title: 'BayesCog: A freely available course in Bayesian statistics and hierarchical Bayesian modeling for psychological science'
+bibliography: paper_psyarxiv.bib
+csl: https://raw.githubusercontent.com/citation-style-language/styles/master/nature.csl
+documentclass: article
+classoption: twocolumn
+geometry: margin=1in
+link-citations: true
+header-includes: |
+  \usepackage{xcolor}
+  \usepackage{fancyhdr}
+  \usepackage{calc}
+  \usepackage{hyperref}
+  \hypersetup{
+    colorlinks=true,
+    linkcolor=blue,
+    citecolor=blue,
+    urlcolor=blue
+  }
+  \setlength{\columnsep}{0.5in}
+  \makeatletter
+  \renewcommand\section{\@startsection {section}{1}{\z@}%
+    {-2ex \@plus -1ex \@minus -.2ex}%
+    {0.5ex \@plus.2ex}%
+    {\normalfont\Large\bfseries}}
+  \renewcommand\subsection{\@startsection{subsection}{2}{\z@}%
+    {-1.5ex\@plus -1ex \@minus -.2ex}%
+    {0.3ex \@plus .2ex}%
+    {\normalfont\large\bfseries}}
+  \def\@maketitle{%
+    \newpage
+    \null
+    \vskip -10em%
+    \begin{center}%
+    \let \footnote \thanks
+      {\LARGE\bfseries \@title \par}%
+      \vskip 1.5em%
+      {\Large Lei Zhang\textsuperscript{1,2,3*} and Aamir Sohail\textsuperscript{1,2*}\par}%
+      \vskip 1.5em%
+      {\small
+        \textsuperscript{1}Centre for Human Brain Health, School of Psychology, University of Birmingham, Birmingham, UK\\
+        \textsuperscript{2}Institute for Mental Health, School of Psychology, University of Birmingham, Birmingham, UK\\
+        \textsuperscript{3}Centre for Developmental Science, School of Psychology, University of Birmingham, Birmingham, UK\par}%
+      \vskip 1em%
+      {\small\textsuperscript{*}Corresponding authors: \{l.zhang.13, axs2210\}[at]bham.ac.uk\par}%
+      \vskip 1em%
+      {\small\textcolor{red}{\textit{This paper is currently under review at the Journal of Open Source Education.}}\par}%
+    \end{center}%
+    \par
+    \vskip 1em}
+  \makeatother
+  \fancypagestyle{firstpage}{%
+    \fancyhf{}
+    \fancyfoot[L]{\footnotesize Corresponding authors: {{l.zhang.13, axs2210}}[at]bham.ac.uk}
+    \renewcommand{\headrulewidth}{0pt}%
+    \renewcommand{\footrulewidth}{0pt}}
+  \AtBeginDocument{\thispagestyle{firstpage}\pagestyle{plain}}
+  \setlength{\csllabelwidth}{2em}
+  \renewcommand{\CSLLeftMargin}[1]{\parbox[t]{\csllabelwidth}{#1}}
+  \renewcommand{\CSLRightInline}[1]{\parbox[t]{\linewidth - \csllabelwidth}{#1}\break}
+---
+
+# Summary
+
[…]
