### 2026-06-04 #376 @editorialbot
[REVIEW]: BayesCog: A freely available course in Bayesian statistics and hierarchical Bayesian modeling for psychological science
**Submitting author:** @sohaamir (<a href="http://orcid.org/0009-0000-6584-4579">Aamir Sohail</a>)
**Repository:** https://github.com/Alpn-Lab/BayesCog
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
**Archive:** 10.31234/osf.io/ua5ng_v1
