### 2026-06-13 ddb4931
Address peer-review comments (website + paper)
WS8 model comparison (R2.8): replace WAIC with PSIS-LOO-CV throughout
(intro, prose, code, master script); remove the now-orphaned
ic_comparison.png figure.

Learning objectives (R1.8): standardised "By the end of this workshop,
you will be able to:" blocks across all workshop intros.

Licensing (R1.2): add MIT LICENSE-MIT for the R/Stan code alongside the
existing CC-BY-SA-4.0 LICENSE for content; document the split in README.

Docs (R1.1): add a "Getting help" section to README and track
CONTRIBUTING.md.

Paper (R1.9): fix "CONTIBUTORS" -> "CONT
[…]
