+++ paper.md
@@ -78,7 +78,7 @@ All course workshops are accompanied by example data and scripts, the latter pro
 
 The BayesCog course provides a general introduction to Bayesian statistics and computational modeling within the context of psychological research. Subsequently, the materials and topics could easily be developed further. This includes using computational methods with neuroimaging data [@Glascher2010; @deHollander2016], understanding behaviors in the social world [@Kutlikova2023; @Pan2023], and the theory-based modeling of psychiatric disorders [@Maia2017; @Sohail2024; @Suter2025]. Furthermore, the Stan programming language remains technically challenging, leading to the development of user-friendly packages for computational modeling [@Ahn2017]. Tutorials on how to implement these packages could further broaden the use of computational methods within the psychological sciences. In the era of large language models (LLMs) and generative AI, the audience are also encouraged to combine this course with LLM tools such as ChatGPT to explore a more tailored learning experience [@Sohail2025a; @Sohail2025b]. 
 
-To this end, we openly receive feedback and suggestions from the wider community. Any tutorial can be adopted, transformed, or built upon for other educational purposes (e.g., courses, single class sessions, workshops) under a Creative Commons Attribution-ShareAlike 4.0 International License. Broader comments can be communicated on the [GitHub repository](https://github.com/Alpn-Lab/BayesCog/issues) by either reporting an issue or requesting an enhancement. On-the-other-hand, we recommend major changes to be made communicated beforehand and – if appropriate - made directly by forking the repository and pushing changes to the main branch. A member of the contributing team will then review the changes. Accepted contributions will be credited and following the community guidelines outlined on the CONTIBUTORS page.
+To this end, we openly receive feedback and suggestions from the wider community. Any tutorial can be adopted, transformed, or built upon for other educational purposes (e.g., courses, single class sessions, workshops) under a Creative Commons Attribution-ShareAlike 4.0 International License. Broader comments can be communicated on the [GitHub repository](https://github.com/Alpn-Lab/BayesCog/issues) by either reporting an issue or requesting an enhancement. On-the-other-hand, we recommend major changes to be made communicated beforehand and – if appropriate - made directly by forking the repository and pushing changes to the main branch. A member of the contributing team will then review the changes. Accepted contributions will be credited and following the community guidelines outlined on the CONTRIBUTING page.
 
 # Acknowledgments
 
