+++ workshops/02.binomial_globe/qmd/mcmc.qmd
@@ -3,29 +3,29 @@ title: "Markov chain Monte Carlo"
 subtitle: "Sampling from the posterior with brute force"
 ---
 
-In our previous chapter, we explored how to estimate the posterior distribution for a simple binomial model using grid approximation. 
-This worked well mainly because we only had a single parameter, $\theta$ to estimate at 20 points in the grid. 
+In our previous chapter, we explored how to estimate the posterior distribution for a simple binomial model using grid approximation.
+This worked well mainly because we only had a single parameter, $\theta$ to estimate at 20 points in the grid.
 
 But what happens when our models become more complex and involve multiple parameters?
 
-Let's consider how the computational challenge scales with additional parameters. If we used 100 grid points for $\theta$, we would need to compute 100 values. 
+Let's consider how the computational challenge scales with additional parameters. If we used 100 grid points for $\theta$, we would need to compute 100 values.
 However, with two parameters $\theta_1$ and $\theta_2$, calculating the marginal likelihood becomes a double integral:
 
 $$p(data) = \int\int p(data,\theta_1,\theta_2)d\theta_1d\theta_2$$
-**This complexity grows dramatically when we move to models with many parameters.** Consider a model with 100 means $(μ)$ and 100 standard deviations $(σ)$. 
+**This complexity grows dramatically when we move to models with many parameters.** Consider a model with 100 means $(μ)$ and 100 standard deviations $(σ)$.
 
 The marginal likelihood becomes:
 
 $$
-\scriptsize 
-p(data) = \int_{\mu_1}\int_{\sigma_1}...\int_{\mu_{100}}\int_{\sigma_{100}} 
-p(data|\mu_1,\sigma_1,...,\mu_{100},\sigma_{100}) \times 
-p(\mu_1,\sigma_1,...,\mu_{100},\sigma_{100}) 
+\scriptsize
+p(data) = \int_{\mu_1}\int_{\sigma_1}...\int_{\mu_{100}}\int_{\sigma_{100}}
+p(data|\mu_1,\sigma_1,...,\mu_{100},\sigma_{100}) \times
+p(\mu_1,\sigma_1,...,\mu_{100},\sigma_{100})
 d\mu_1d\sigma_1...d\mu_{100}d\sigma_{100}
 $$
 If we use 100 grid points for each parameter in this model, the computational burden grows exponentially. Even with just two parameters, we need 100 × 100 = 10,000 computations, and with five parameters, this becomes 100⁵ = 10 billion computations!
 
-Fortunately, there exists a solution: Markov chain Monte Carlo (MCMC). 
+Fortunately, there exists a solution: Markov chain Monte Carlo (MCMC).
 
 **Instead of trying to compute the entire posterior distribution, MCMC provides a way to draw samples; approximating it.** MCMC takes advantage of the proportionality:
 
@@ -38,7 +38,7 @@ Before we discuss an example of how MCMC works to sample from the posterior, let
 
 ### Markov chains
 
-A Markov chain is a sequence of random states where the probability of transitioning to the next state depends only on the current state, and not on the sequence of states that preceded it. 
+A Markov chain is a sequence of random states where the probability of transitioning to the next state depends only on the current state, and not on the sequence of states that preceded it.
 
 Mathematically, for a sequence of random variables $X_1, X_2, ..., X_n$, the Markov property states:
 
@@ -87,7 +87,7 @@ In general, here are the steps behind MCMC:
 
 ### A practical example of MCMC using the Metropolis algorithm
 
-Here's an practical example demonstrating the idea behind MCMC, using a specific instance, the Metropolis algorithm. 
[…]

[…]
