### 2026-06-15 9373a5d
Address peer-review comments (sessions 3–5)
Website / workshops:
- WS2 (binomial_globe): add real MCMC rank-ECDF plot (R2.7), recoloured to
  match stan_trace palette; wire generation into student + master scripts
- WS6 (reinforcement_learning): truncated Normal vs Beta(2,5) prior figure +
  rewritten "Setting priors" callout (R2.3); "belief"->"assumption" reword
- WS1 (R_basics): "keep learning R" callout + free resource links (R1.5)
- rl_concepts: prose typo fix
- course_overview / README: remove WIP workshop rows (R1.3)
- renv.lock: record bayesplot + deps

Manuscripts (paper.md + paper_psy
[…]
