+++ paper.md
@@ -50,9 +50,11 @@ P(\theta|D) = \frac{P(D|\theta) \cdot P(\theta)}{P(D)}
 
 where $P(\theta|D)$ is the posterior distribution, $P(D|\theta)$ is the likelihood, $P(\theta)$ is the prior distribution, and $P(D)$ is the marginal likelihood.
 
-Bayesian methods confer advantages over frequentist approaches (e.g., Maximum Likelihood Estimation, MLE), by quantifying the uncertainty, and when implemented hierarchically, permit simultaneous estimation of individual and group-level parameters while appropriately pooling information across participants [@Lee2011]. Historically restrictive due to their computational burden, these methods are now more accessible though the development of multiple programming languages and software such as JAGS [@Plummer2003] and Stan [@Carpenter2017] which optimize the sampling process used for parameter estimation using approaches such as Markov chain Monte Carlo (MCMC). 
+Bayesian methods confer advantages over frequentist approaches (e.g., Maximum Likelihood Estimation, MLE), by quantifying the uncertainty, and when implemented hierarchically, permit simultaneous estimation of individual and group-level parameters while appropriately pooling information across participants [@Lee2011]. Historically restrictive due to their computational burden, these methods are now more accessible though the development of multiple programming languages and software such as JAGS [@Plummer2003] and Stan [@Carpenter2017] which optimize the sampling process used for parameter estimation using approaches such as Markov chain Monte Carlo (MCMC).
 
 Using Bayesian models of cognition in one’s own research requires a conceptual understanding of both Bayesian statistics and cognitive modeling, as well as the practical skills to translate these models into computer code. Textbooks [@Kruschke2014; @Lambert2018; @Lee2014; @McElreath2018] and tutorial papers [@Baribault2023; @Lockwood2021; @Wilson2019; @Zhang2020] have made learning these skills more accessible, but are often not freely available, and challenging for researchers, especially early career researchers with little to no prior experience. Additionally, whilst free, online courses for the computational modeling of cognition currently exist, these are few and far between, and exclusively cover non-Bayesian implementations in Python [@Rhoads2022] and MATLAB [@OReilly2015], as well as Bayesian approaches in Python [@NivLab2021]. A full course implementing Bayesian models of cognition through the open source R programming language is therefore a valuable yet currently non-existent resource.
+
+BayesCog was originally developed by Lei Zhang at UKE Hamburg as a short course in Bayesian statistics and cognitive modeling for PhD students. The course was subsequently expanded by Lei at the University of Vienna, where the course received a Commendation Award from the Society for the Improvement of Psychological Science (SIPS), as well as an Early Career Teaching Award by the University. To make these materials more widely and openly available, Aamir Sohail, a PhD student working with Lei at the University of Birmingham, subsequently converted, edited, and expanded the source material into the present online course. Both authors are active researchers in computational modeling and cognitive neuroscience who have applied these Bayesian methods in their own empirical work which investigates the cognitive processes of learning and decision making in social contexts [@Zhang2020; @Crawley2020; @ZhangGlascher2020; @Sohail2026].
 
 # Learning Objectives
 
@@ -64,30 +66,51 @@ In the course, students will:
 *	Learn basic reinforcement learning (RL) concepts and build a simple RL model (Rescorla-Wagner) using RStan
 *	Be able to evaluate model performance through model comparison, parameter recovery and posterior predictive checks
 
[…]
