+++ paper_psyarxiv.md
@@ -82,6 +82,8 @@ Bayesian methods confer advantages over frequentist approaches (e.g., Maximum Li
 
 Using Bayesian models of cognition in one’s own research requires a conceptual understanding of both Bayesian statistics and cognitive modeling, as well as the practical skills to translate these models into computer code. Textbooks [@Kruschke2014; @Lambert2018; @Lee2014; @McElreath2018] and tutorial papers [@Baribault2023; @Lockwood2021; @Wilson2019; @Zhang2020] have made learning these skills more accessible, but are often not freely available, and challenging for researchers, especially early career researchers with little to no prior experience. Additionally, whilst free, online courses for the computational modeling of cognition currently exist, these are few and far between, and exclusively cover non-Bayesian implementations in Python [@Rhoads2022] and MATLAB [@OReilly2015], as well as Bayesian approaches in Python [@NivLab2021]. A full course implementing Bayesian models of cognition through the open source R programming language is therefore a valuable yet currently non-existent resource.
 
+BayesCog was originally developed by Lei Zhang at UKE Hamburg as a short course in Bayesian statistics and cognitive modeling for PhD students. The course was subsequently expanded by Lei at the University of Vienna, where the course received a Commendation Award from the Society for the Improvement of Psychological Science (SIPS), as well as an Early Career Teaching Award by the University. To make these materials more widely and openly available, Aamir Sohail, a PhD student working with Lei at the University of Birmingham, subsequently converted, edited, and expanded the source material into the present online course. Both authors are active researchers in computational modeling and cognitive neuroscience who have applied these Bayesian methods in their own empirical work which investigates the cognitive processes of learning and decision making in social contexts [@Zhang2020; @Crawley2020; @ZhangGlascher2020; @Sohail2026].
+
 # Learning Objectives
 
 In the course, students will:
@@ -94,7 +96,40 @@ In the course, students will:
 -   Learn basic reinforcement learning (RL) concepts and build a simple RL model (Rescorla-Wagner) using RStan
 -   Be able to evaluate model performance through model comparison, parameter recovery and posterior predictive checks
 
-Prior to the first workshop, the course begins by summarizing the broader philosophy in which the techniques and methods are implemented. Specifically, this concerns Marr’s influential three levels of analysis [@Marr1982], which describe how algorithmic-level (as opposed to computational and implementation levels) models can help understand behaviour. Doing so shapes the course material within this framework, demonstrating the necessity for building strong theories in psychology [@Press2022]. Assuming no prior experience with programming, the course properly (Workshop 1) begins with a basic introduction to the R programming language and the RStudio interface [@RStudio2020]. This introductory workshop first provides a general introduction to data structures, variables, and packages, after which students will work with simulated data from a reversal learning task, performing basic summary statistics including correlation and regression, and visualizing the results using the `ggplot2` package [@Wickham2016].
+\begin{table*}[t]
+\centering
+\small
+\caption{Conceptual foundations of the BayesCog course (Weeks 1--3).}
+\begin{tabular}{p{1cm}p{2.2cm}p{4cm}p{7.3cm}}
+\hline
+\textbf{Week} & \textbf{Workshop} & \textbf{Topic} & \textbf{Focus} \\
+\hline
+1 & Workshop 1 & Introduction to R and RStudio & Foundational programming in R: data structures, packages, summary statistics and visualization \\
+2 & Workshop 2 & Introduction to Bayesian statistics & Bayesian versus frequentist definitions of probability \\
[…]
