### 2026-06-16 7d73cc7
Add MIT-license note for code to both papers (R1.2 follow-up)
Append a clause to the CC BY-SA sentence in Future Directions noting that
the accompanying code (R scripts and Stan models) is released under the MIT
License, documenting the content/code dual-license split. Rebuild both PDFs.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
