### 2026-06-16 ead8e36
Relicense repository under a single MIT license
Replace the dual CC BY-SA 4.0 (content) + MIT (code) setup with a single
OSI-approved MIT license covering the whole repository, per reviewer R1.2.

- LICENSE: CC BY-SA 4.0 text -> MIT (copyright 2020-2026 Lei Zhang and
  Aamir Sohail)
- Remove the now-redundant LICENSE-MIT
- README and website (index.qmd) License sections rewritten to single MIT;
  drop the CC BY-SA badge image

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
