### 2026-06-16 d1bdcb5
Update author affiliations and paper prose; rebuild PDFs
- Add affiliations for L.Z. (University of Vienna; UKE Hamburg) and A.S.
  (Centre for Developmental Science); add B15 2TT postcodes
- Refine Summary, Statement of Need, development-history (L.Z.), and the
  delivery paragraph; shorten the Table 1/2 captions
- Switch the Future Directions license sentence to single MIT
- Mirror all prose edits into paper_psyarxiv.md and rebuild both PDFs
  (paper.pdf via inara, paper_psyarxiv.pdf via quarto pandoc)

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
