+++ paper.md
@@ -17,30 +17,36 @@ tags:
 authors:
   - name: Lei Zhang
     orcid: 0000-0002-9586-595X
-    affiliation: "1, 2, 3"
+    affiliation: "1, 2, 3, 4, 5, 6"
   - name: Aamir Sohail
     orcid: 0009-0000-6584-4579
-    affiliation: "1, 2"
+    affiliation: "1, 2, 3"
 affiliations:
- - name: Centre for Human Brain Health, School of Psychology, University of Birmingham, Birmingham, UK
+ - name: School of Psychology, University of Birmingham, Birmingham, B15 2TT, UK
    index: 1
- - name: Institute for Mental Health, School of Psychology, University of Birmingham, Birmingham, UK
+ - name: Centre for Human Brain Health, University of Birmingham, Birmingham, B15 2TT, UK
    index: 2
- - name: Centre for Developmental Science, School of Psychology, University of Birmingham, Birmingham, UK
+ - name: Institute for Mental Health, University of Birmingham, Birmingham, B15 2TT, UK
    index: 3
+ - name: Centre for Developmental Science, University of Birmingham, B15 2TT, Birmingham, UK
+   index: 4
+ - name: Social, Cognitive and Affective Neuroscience Unit, Department of Cognition, Emotion, and Methods, Faculty of Psychology, University of Vienna, Vienna, 1010, Austria
+   index: 5
+ - name: Institute of Systems Neuroscience, University Medical Center Hamburg-Eppendorf, Hamburg, 20246, Germany
+   index: 6
 date: 2025-07-12
 bibliography: paper.bib
 ---
 
 # Summary
 
-We present [BayesCog](https://alpn-lab.github.io/BayesCog/), an openly-available online course for the computational modeling of human behavior (i.e., cognitive modeling) using Bayesian inference, with reinforcement learning as a core example throughout the course. Assuming little to no prior experience, audience of this course  will be formally grounded in key concepts including Bayesian statistics and reinforcement learning, and practically, will build, assess, compare, and validate models using the R interface to the Stan programming language, RStan. Starting with binary choice models, the audience will learn to estimate parameters representing latent components of behavior by fitting reinforcement learning models, both at the individual and group-level, eventually with hierarchical modeling.
+We present [BayesCog](https://alpn-lab.github.io/BayesCog/), an openly-available online course for the computational modeling of human behavior (i.e., cognitive modeling) using Bayesian inference, with reinforcement learning as a core example throughout the course. Assuming little to no prior experience, users of this course will be formally grounded in key concepts including Bayesian statistics and reinforcement learning, and practically, will build, assess, compare, and validate models using the R interface to the Stan programming language, RStan. Starting with binary choice models, users will learn to estimate parameters representing latent components of behavior by fitting reinforcement learning models, both at the individual and group-level, and eventually with hierarchical structures.
 
 The course is generally suitable for those interested in developing models of human cognition at any level of experience. In making the course openly available, we aim for computational modeling under the Bayesian approach to be more strongly represented in the psychological sciences.
 
 # Statement of Need
 
[…]
