+++ paper_psyarxiv.md
@@ -35,12 +35,15 @@ header-includes: |
     \let \footnote \thanks
       {\LARGE\bfseries \@title \par}%
       \vskip 1.5em%
-      {\Large Lei Zhang\textsuperscript{1,2,3*} and Aamir Sohail\textsuperscript{1,2*}\par}%
+      {\Large Lei Zhang\textsuperscript{1,2,3,4,5,6*} and Aamir Sohail\textsuperscript{1,2,3*}\par}%
       \vskip 1.5em%
       {\small
-        \textsuperscript{1}Centre for Human Brain Health, School of Psychology, University of Birmingham, Birmingham, UK\\
-        \textsuperscript{2}Institute for Mental Health, School of Psychology, University of Birmingham, Birmingham, UK\\
-        \textsuperscript{3}Centre for Developmental Science, School of Psychology, University of Birmingham, Birmingham, UK\par}%
+        \textsuperscript{1}School of Psychology, University of Birmingham, Birmingham, B15 2TT, UK\\
+        \textsuperscript{2}Centre for Human Brain Health, University of Birmingham, Birmingham, B15 2TT, UK\\
+        \textsuperscript{3}Institute for Mental Health, University of Birmingham, Birmingham, B15 2TT, UK\\
+        \textsuperscript{4}Centre for Developmental Science, University of Birmingham, B15 2TT, Birmingham, UK\\
+        \textsuperscript{5}Social, Cognitive and Affective Neuroscience Unit, Department of Cognition, Emotion, and Methods, Faculty of Psychology, University of Vienna, Vienna, 1010, Austria\\
+        \textsuperscript{6}Institute of Systems Neuroscience, University Medical Center Hamburg-Eppendorf, Hamburg, 20246, Germany\par}%
       \vskip 1em%
       {\small\textsuperscript{*}Corresponding authors: \{l.zhang.13, axs2210\}[at]bham.ac.uk\par}%
       \vskip 1em%
@@ -62,13 +65,13 @@ header-includes: |
 
 # Summary
 
-We present \textbf{BayesCog} (\texttt{\url{https://alpn-lab.github.io/BayesCog}}), an openly-available online course for the computational modeling of human behaviour (i.e., cognitive modeling) using Bayesian inference, with reinforcement learning as a core example throughout the course. Assuming little to no prior experience, audience of this course will be formally grounded in key concepts including Bayesian statistics and reinforcement learning, and practically, will build, assess, compare, and validate models using the R interface to the Stan programming language, RStan. Starting with binary choice models, the audience will learn to estimate parameters representing latent components of behaviour by fitting reinforcement learning models, both at the individual and group-level, eventually with hierarchical modeling.
+We present \textbf{BayesCog} (\texttt{\url{https://alpn-lab.github.io/BayesCog}}), an openly-available online course for the computational modeling of human behaviour (i.e., cognitive modeling) using Bayesian inference, with reinforcement learning as a core example throughout the course. Assuming little to no prior experience, users of this course will be formally grounded in key concepts including Bayesian statistics and reinforcement learning, and practically, will build, assess, compare, and validate models using the R interface to the Stan programming language, RStan. Starting with binary choice models, users will learn to estimate parameters representing latent components of behaviour by fitting reinforcement learning models, both at the individual and group-level, and eventually with hierarchical structures.
 
 The course is generally suitable for those interested in developing models of human cognition at any level of experience. In making the course openly available, we aim for computational modeling under the Bayesian approach to be more strongly represented in the psychological sciences.
 
 # Statement of Need
 
[…]
