### 2026-06-16 da7c71d
Add Dependabot config for GitHub Actions updates
Track action versions in .github/workflows/ (weekly, assigned to sohaamir).
R/Docker/Python environments left untracked to preserve reproducibility.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
