### 2026-06-16 a89fbf6
ci: bump EndBug/add-and-commit from 9 to 10
Bumps [EndBug/add-and-commit](https://github.com/endbug/add-and-commit) from 9 to 10.
- [Release notes](https://github.com/endbug/add-and-commit/releases)
- [Commits](https://github.com/endbug/add-and-commit/compare/v9...v10)

---
updated-dependencies:
- dependency-name: EndBug/add-and-commit
  dependency-version: '10'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
