### 2026-06-21 17a3281
ci: skip PDF commit-back step on Dependabot PRs
Dependabot-triggered runs get a read-only GITHUB_TOKEN, so the
add-and-commit push fails and reports the whole Draft PDF check as
failed even though the paper builds fine. Guard the commit step on the
actor so bot PRs stay green while keeping the build signal.

Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>
