+++ .github/workflows/draft-pdf.yml
@@ -29,6 +29,9 @@ jobs:
           # paper.md
           path: paper.pdf
       - name: Commit PDF to repository
+        # Dependabot PRs run with a read-only token and cannot push, so skip the
+        # commit-back step for them (the PDF still builds, keeping that signal).
+        if: ${{ github.actor != 'dependabot[bot]' }}
         uses: EndBug/add-and-commit@v10
         with:
             message: '(auto) Paper PDF Draft'
