### 2026-08-20 #376 @sohaamir
@magsol the problematic references with missing/skipped DOIs do not have DOIs to include:

- The 'NivStan' and 'Models of Learning' citations are websites
- 'The JAGS' and 'RStudio' citations are software with no DOIs
- The citation to 'A theory of Pavlovian conditioning: Variations in the effectiveness of reinforcement and non-reinforcement' does not have a DOI

I am happy to proceed how you best see fit. I could remove these citations entirely or find suitable replacements?
