### 2026-08-20 #376 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
