### 2026-08-21 #376 @sohaamir
Both myself and @lei-zhang are happy with the paper PDF, is there anything needed from our end?
