### 2026-08-24 #376 @magsol
@sohaamir With the `recommend-accept` tag in the previous post 4 days ago, we're now waiting for the JOSE EiCs to come in and give it a final once-over. Once they've had a chance to do that and approve of it, it'll be published.

Unfortunately they're busy folks and we don't have a time frame for when that'll happen. I'll ping them behind the scenes, but it's just a matter of waiting until they have a moment. I appreciate your patience! Almost there!
